+++ README.md
@@ -5,7 +5,6 @@
 <br>
 
 - 배포 url: https://makeup-product.vercel.app/
-  
 - 상품 타입, 카테고리, 태그 별로 상품을 찾아볼 수 있습니다.
 
 <br>
@@ -29,9 +28,11 @@ Package Manager: `Yarn`
 - **헤더**
 
 - **GNB (Global Navigation Bar)**
+
   - 상품 타입
 
 - **SNB (Side Navigation Bar)**
+
   - 상품 필터 ( 카테고리, 태그 )
 
 - **본문**
@@ -42,10 +43,12 @@ Package Manager: `Yarn`
 ### 2단계: API 연동 및 데이터 표시
 
 - **조회 API 연동**
+
   - 전체 상품 리스트
   - 상품 필터링된 리스트 ( 타입별, 카테고리별, 태그별 )
 
 - **Skeleton UI 추가**
+
   - 상품이 많아 렌더링되는데 시간이 오래걸리는 문제가 발생하여 사용자 경험 향상을 위해 적용
 
 - **Floating Up Button 추가**
@@ -54,17 +57,32 @@ Package Manager: `Yarn`
 
 <br>
 
+### 3단계: 반응형 디자인 구현
+
+- 반응형 디자인 구현
+
+  - 헤더
+  - GNB
+  - SNB
+  - 본문
+
+- px를 없애고, media query 적용
+
+<br>
+
 ## 프로젝트 실행 방법
 
 1. Env 세팅
-root directory
+   root directory
+
 ```shell
-NEXT_PUBLIC_API = 
+NEXT_PUBLIC_API =
 ```
 
 <br>
 
 2. 프로젝트 실행
+
 ```shell
 # 필요한 모듈 설치
 yarn
