+++ next.config.mjs
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: false,
+  reactStrictMode: true,
+  distDir: 'build', // 정적 웹 사이트 구축을 위한 설정
   // output: 'export', // 정적 웹 사이트 구축을 위한 설정
   trailingSlash: true, // 정적 웹 사이트 구축을 위한 설정
   images: {
@@ -19,27 +20,6 @@ const nextConfig = {
         hostname: 'montevista.greatheartsamerica.org',
       },
     ],
-  },
-  async headers() {
-    return [
-      {
-        // matching all API routes
-        source: '/api/:path*',
-        headers: [
-          { key: 'Access-Control-Allow-Credentials', value: 'true' },
-          { key: 'Access-Control-Allow-Origin', value: '*' }, // replace this your actual origin
-          {
-            key: 'Access-Control-Allow-Methods',
-            value: 'GET,DELETE,PATCH,POST,PUT',
-          },
-          {
-            key: 'Access-Control-Allow-Headers',
-            value:
-              'X-CSRF-Token, X-Requested-With, Accept, Accept-Version, Content-Length, Content-MD5, Content-Type, Date, X-Api-Version',
-          },
-        ],
-      },
-    ];
   },
 };
